feat(construction): Split up construction module into several
I have come to realize that the existing construction module was already 1400 lines long and we are far from finished adding to it.  I'll probably implement window frame objects as part of the WindowConstruction and I think DynamicWindowConstruction is going to be it's own object, which holds ShadeControl objects. Given all of this, it will be a lot easier to maintain if we split up the construction module now before we build more on top of it.
