+++ honeybee_energy/properties/door.py
@@ -1,6 +1,7 @@
 # coding=utf-8
 """Door Energy Properties."""
-from ..construction import OpaqueConstruction, WindowConstruction
+from ..construction.opaque import OpaqueConstruction
+from ..construction.window import WindowConstruction
 from ..lib.constructionsets import generic_construction_set
 
 
@@ -8,8 +9,8 @@ class DoorEnergyProperties(object):
     """Energy Properties for Honeybee Door.
 
     Properties:
-        construction
-        is_construction_set_by_user
+        * construction
+        * is_construction_set_by_user
     """
 
     __slots__ = ('_host', '_construction')
@@ -72,7 +73,8 @@ class DoorEnergyProperties(object):
 
     @property
     def is_construction_set_by_user(self):
-        """Check if construction is set by user."""
+        """Boolean noting if construction is user-set (as opposed to a ConstructionSet).
+        """
         return self._construction is not None
 
     @classmethod
