+++ honeybee_energy/properties/face.py
@@ -1,6 +1,6 @@
 # coding=utf-8
 """Face Energy Properties."""
-from ..construction import OpaqueConstruction
+from ..construction.opaque import OpaqueConstruction
 from ..lib.constructionsets import generic_construction_set
 
 
@@ -8,8 +8,8 @@ class FaceEnergyProperties(object):
     """Energy Properties for Honeybee Face.
 
     Properties:
-        construction
-        is_construction_set_by_user
+        * construction
+        * is_construction_set_by_user
     """
 
     __slots__ = ('_host', '_construction')
@@ -61,7 +61,8 @@ class FaceEnergyProperties(object):
 
     @property
     def is_construction_set_by_user(self):
-        """Check if construction is set by user."""
+        """Boolean noting if construction is user-set (as opposed to a ConstructionSet).
+        """
         return self._construction is not None
 
     @classmethod
